refactor(cli): extract attachment commands to bin/commands/attachments.js
Move `attachments`, `attachment-upload`, and `attachment-delete` into a
dedicated module. The download path's inline `sanitizeFilename` is
duplicated here rather than imported, keeping the module self-contained
(the export command keeps its own copy when it moves out next).
